+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.data.TaskList;
+import duke.ui.Ui;
+import duke.storage.Storage;
+
 public abstract class Command {
     public Command() {
     }
