+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,7 +1,13 @@
-public class DoneCommand extends Command {
+package duke.commands;
+
+import duke.data.TaskList;
+import duke.ui.Ui;
+import duke.storage.Storage;
+
+public class DeleteCommand extends Command {
     private final int taskNumber;
 
-    public DoneCommand(String userInput) {
+    public DeleteCommand(String userInput) {
         this.taskNumber = Integer.parseInt(
                 userInput.replaceAll(
                         "[^0-9]",
@@ -10,7 +16,7 @@ public class DoneCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.printMsg(tasks.markAsDone(this.taskNumber));
+        ui.printMsg(tasks.deleteTask(this.taskNumber));
         storage.write(tasks.getSaveData());
     }
 }
