+++ src/main/java/duke/data/exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.data.exception;
+
 public class DukeException extends RuntimeException {
     public DukeException(String message) {
         super("! " + message + " !");
