+++ src/main/java/duke/data/tasks/Deadlines.java
@@ -1,3 +1,5 @@
+package duke.data.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
