+++ src/main/java/duke/data/tasks/ToDos.java
@@ -1,3 +1,5 @@
+package duke.data.tasks;
+
 public class ToDos extends Task {
     public ToDos(String name) {
         super(name);
