+++ src/main/java/duke/ui/Ui.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 import java.io.InputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
